Address comments and put shape operations in their own block
