+++ ndonnx/_experimental.py
@@ -1,7 +1,7 @@
 # Copyright (c) QuantCo 2023-2024
 # SPDX-License-Identifier: BSD-3-Clause
 
-from ._core import OperationsBlock
+from ._core import OperationsBlock, UniformShapeOperations
 from ._data_types import CastMixin, Schema, StructType
 
 __all__ = [
@@ -9,4 +9,5 @@ __all__ = [
     "Schema",
     "CastMixin",
     "OperationsBlock",
+    "UniformShapeOperations",
 ]
